extract model converter code from mojo
